% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_afstemningsomraader.R
\docType{data}
\name{geo_afstemningsomraader}
\alias{geo_afstemningsomraader}
\title{Data from the DAGI API}
\format{
A data frame with 1383 rows and 24 variables:
\describe{
  \item{afstemningssted_adgangspunkt_x}{see DAGI documentation}
  \item{afstemningssted_adgangspunkt_y}{see DAGI documentation}
  \item{afstemningsstedadressebetegnelse}{see DAGI documentation}
  \item{afstemningsstedadresseid}{see DAGI documentation}
  \item{afstemningsstednavn}{see DAGI documentation}
  \item{changed}{see DAGI documentation}
  \item{dagi_id}{see DAGI documentation}
  \item{geo_changed}{see DAGI documentation}
  \item{geo_version}{see DAGI documentation}
  \item{geometry}{SF geometry}
  \item{kommunekode}{see DAGI documentation}
  \item{kommunenavn}{see DAGI documentation}
  \item{navn}{see DAGI documentation}
  \item{nummer}{see DAGI documentation}
  \item{opstillingskredsnavn}{see DAGI documentation}
  \item{opstillingskredsnummer}{see DAGI documentation}
  \item{regionskode}{see DAGI documentation}
  \item{regionsnavn}{see DAGI documentation}
  \item{storkredsnavn}{see DAGI documentation}
  \item{storkredsnummer}{see DAGI documentation}
  \item{valglandsdelsbogstav}{see DAGI documentation}
  \item{valglandsdelsnavn}{see DAGI documentation}
  \item{visueltcenter_x}{see DAGI documentation}
  \item{visueltcenter_y}{see DAGI documentation}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(geo_afstemningsomraader)
}
\description{
Data from the DAGI API
}
\examples{
data(geo_afstemningsomraader)


}
\keyword{datasets}
