% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_postnumre.R
\docType{data}
\name{geo_postnumre}
\alias{geo_postnumre}
\title{Data from the DAGI API}
\format{
A data frame with 1069 rows and 10 variables:
\describe{
  \item{nr}{see DAGI documentation}
  \item{navn}{see DAGI documentation}
  \item{stormodtager}{see DAGI documentation}
  \item{visueltcenter_x}{see DAGI documentation}
  \item{visueltcenter_y}{see DAGI documentation}
  \item{changed}{see DAGI documentation}
  \item{geo_changed}{see DAGI documentation}
  \item{geo_version}{see DAGI documentation}
  \item{dagi_id}{see DAGI documentation}
  \item{geometry}{SF geometry}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(geo_postnumre)
}
\description{
Data from the DAGI API
}
\examples{
data(geo_postnumre)


}
\keyword{datasets}
