% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_regioner.R
\docType{data}
\name{geo_regioner}
\alias{geo_regioner}
\title{Data from the DAGI API}
\format{
A data frame with 5 rows and 16 variables:
\describe{
  \item{dagi_id}{see DAGI documentation}
  \item{kode}{see DAGI documentation}
  \item{navn}{see DAGI documentation}
  \item{nuts2}{see DAGI documentation}
  \item{changed}{see DAGI documentation}
  \item{geo_changed}{see DAGI documentation}
  \item{geo_version}{see DAGI documentation}
  \item{visueltcenter_x}{see DAGI documentation}
  \item{visueltcenter_y}{see DAGI documentation}
  \item{population}{number of inhabitants in region from DST}
  \item{men}{number of men in region from DST}
  \item{women}{number of women in region from DST}
  \item{pct_men}{percentage of men in region from DST}
  \item{avg_age}{Average age in region from DST}
  \item{stat_year}{Year of stats from DST}
  \item{geometry}{SF geometry}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(geo_regioner)
}
\description{
Data from the DAGI API
}
\examples{
data(geo_regioner)


}
\keyword{datasets}
