% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kom_tabel.R
\docType{data}
\name{kom_tabel}
\alias{kom_tabel}
\title{Data to change data from old county codes to new ones, from 2007}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{X}{x coordinate for centroid of old county}
  \item{Y}{y coordinate for centroid of old county}
  \item{kode_f07}{code for county before 2007}
  \item{kommune}{new county after 2007}
  \item{kommune_f07}{name for county before 2007}
}
}
\source{
calculated from centroids of old counties to new
}
\usage{
data(kom_tabel)
}
\description{
Data to change data from old county codes to new ones, from 2007
}
\examples{
data(kom_tabel)


}
\keyword{datasets}
