% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_kommuner.R
\docType{data}
\name{std_kommuner}
\alias{std_kommuner}
\title{Data from DST for age and gender standardization on county level}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{KØN}{gender}
  \item{ALDER}{age}
  \item{INDHOLD}{n}
  \item{area}{area code}
}
}
\source{
\href{https://www.dst.dk/da/}{Website}
}
\usage{
data(std_kommuner)
}
\description{
Data from DST for age and gender standardization on county level
}
\examples{
data(std_kommuner)


}
\keyword{datasets}
