% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_sogne.R
\docType{data}
\name{std_sogne}
\alias{std_sogne}
\title{Data from DST for age and gender standardization on Parrish level}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{KØN}{gender}
  \item{ALDER}{age}
  \item{INDHOLD}{n}
  \item{kode}{parrish code}
}
}
\source{
\href{https://www.dst.dk/da/}{Website}
}
\usage{
data(std_sogne)
}
\description{
Data from DST for age and gender standardization on Parrish level
}
\examples{
data(std_sogne)


}
\keyword{datasets}
